Add 'completed' field for sweep extraction only
